# public/openapi.yml
openapi: 3.0.3
info:
  version: "1"
  title: "Food delivery API"
  description: "Demo Food delivery API"
paths:
...
  /courier:
    post:
      summary: Create
      operationId: createCourier
      tags:
        - Courier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourierRequestBody"
      responses:
        200:
          description: Courier Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourierSchema"
    get:
      summary: Get Couriers
      operationId: getCouriers
      tags:
        - Courier
      parameters:
        - in: query
          name: filter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommonFilter"
      responses:
        200:
          description: Couriers Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouriersResponse"
    delete:
      summary: Delete all Couriers
      tags:
        - Courier
      operationId: deleteAllCouriers
      responses:
        204:
          description: Success
  /courier/{id}:
    patch:
      summary: Update
      operationId: updateCourierById
      tags:
        - Courier
      parameters:
        - name: id
          in: path
          description: Courier ID
          required: true
          schema:
            type: number
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourierRequestBody"
      responses:
        200:
          description: Courier Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourierSchema"
    get:
      summary: Get Courier by ID
      operationId: getCourierById
      tags:
        - Courier
      parameters:
        - name: id
          in: path
          description: Courier ID
          required: true
          schema:
            type: number
          example: 1
      responses:
        200:
          description: Courier Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourierSchema"
    delete:
      summary: Delete Courier by ID
      tags:
        - Courier
      operationId: deleteOneCourier
      parameters:
        - name: id
          in: path
          description: Courier ID
          required: true
          schema:
            type: number
          example: 1
      responses:
        204:
          description: Success
  # END Courier

  # Product
...
  /customer:
    post:
      summary: Create
      operationId: createCustomer
      tags:
        - Customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerRequestBody"
      responses:
        200:
          description: Customer Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerSchema"
    get:
      summary: Get Customers
      operationId: getCustomers
      tags:
        - Customer
      parameters:
        - in: query
          name: filter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommonFilter"
      responses:
        200:
          description: Customers Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomersResponse"
    delete:
      summary: Delete all Customers
      tags:
        - Customer
      operationId: deleteAllCustomers
      responses:
        204:
          description: Success
  /customer/{id}:
    patch:
      summary: Update
      operationId: updateCustomerById
      tags:
        - Customer
      parameters:
        - name: id
          in: path
          description: Customer ID
          required: true
          schema:
            type: number
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerRequestBody"
      responses:
        200:
          description: Customer Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerSchema"
    get:
      summary: Get Customer by ID
      operationId: getCustomerById
      tags:
        - Customer
      parameters:
        - name: id
          in: path
          description: Customer ID
          required: true
          schema:
            type: number
          example: 1
      responses:
        200:
          description: Customer Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerSchema"
    delete:
      summary: Delete Customer by ID
      tags:
        - Customer
      operationId: deleteOneCustomer
      parameters:
        - name: id
          in: path
          description: Customer ID
          required: true
          schema:
            type: number
          example: 1
      responses:
        204:
          description: Success
  # END Customer

  # Order
  /order:
    post:
      summary: Create
      operationId: createOrder
      tags:
        - Order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderRequestBody"
      responses:
        200:
          description: Order Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderSchema"
    get:
      summary: Get Orders
      operationId: getOrders
      tags:
        - Order
      parameters:
        - in: query
          name: filter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommonFilter"
      responses:
        200:
          description: Orders Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersResponse"
    delete:
      summary: Delete all Orders
      tags:
        - Order
      operationId: deleteAllOrders
      responses:
        204:
          description: Success
  /order/{id}:
    patch:
      summary: Update
      operationId: updateOrderById
      tags:
        - Order
      parameters:
        - name: id
          in: path
          description: Order ID
          required: true
          schema:
            type: number
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderRequestBody"
      responses:
        200:
          description: Order Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderSchema"
    get:
      summary: Get Order by ID
      operationId: getOrderById
      tags:
        - Order
      parameters:
        - name: id
          in: path
          description: Order ID
          required: true
          schema:
            type: number
          example: 1
      responses:
        200:
          description: Order Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderSchema"
    delete:
      summary: Delete Order by ID
      tags:
        - Order
      operationId: deleteOneOrder
      parameters:
        - name: id
          in: path
          description: Order ID
          required: true
          schema:
            type: number
          example: 1
      responses:
        204:
          description: Success
  # END Order

  # OrderItem
...
  /courier-order:
    post:
      summary: Create
      operationId: createCourierOrder
      tags:
        - CourierOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourierOrderRequestBody"
      responses:
        200:
          description: CourierOrder Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourierOrderSchema"
    get:
      summary: Get CourierOrders
      operationId: getCourierOrders
      tags:
        - CourierOrder
      parameters:
        - in: query
          name: filter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommonFilter"
      responses:
        200:
          description: CourierOrders Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourierOrdersResponse"
    delete:
      summary: Delete all CourierOrders
      tags:
        - CourierOrder
      operationId: deleteAllCourierOrders
      responses:
        204:
          description: Success
  /courier-order/{id}:
    patch:
      summary: Update
      operationId: updateCourierOrderById
      tags:
        - CourierOrder
      parameters:
        - name: id
          in: path
          description: CourierOrder ID
          required: true
          schema:
            type: number
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourierOrderRequestBody"
      responses:
        200:
          description: CourierOrder Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourierOrderSchema"
    get:
      summary: Get CourierOrder by ID
      operationId: getCourierOrderById
      tags:
        - CourierOrder
      parameters:
        - name: id
          in: path
          description: CourierOrder ID
          required: true
          schema:
            type: number
          example: 1
      responses:
        200:
          description: CourierOrder Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourierOrderSchema"
    delete:
      summary: Delete CourierOrder by ID
      tags:
        - CourierOrder
      operationId: deleteOneCourierOrder
      parameters:
        - name: id
          in: path
          description: CourierOrder ID
          required: true
          schema:
            type: number
          example: 1
      responses:
        204:
          description: Success
  # END CourierOrder
components:
  schemas:
...
    CourierSchema:
      description: Courier schema
      type: object
      required:
        - id
        - name
        - created_at
        - updated_at
      properties:
        id:
          type: number
        name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      example:
        id: 1
        name: Some Courier
        created_at: '2020-02-10 17:32:28'
        updated_at: '2020-02-10 17:32:28'
    CourierRequestBody:
      description: Create Courier
      type: object
      required:
        - name
      properties:
        name:
          type: string
      example:
        name: Some Courier
    CouriersResponse:
      description: List of Couriers
      type: array
      items:
        $ref: '#/components/schemas/CourierSchema'
...
    CustomerSchema:
      description: Customer schema
      type: object
      required:
        - id
        - name
        - address
        - created_at
        - updated_at
      properties:
        id:
          type: number
        name:
          type: string
        address:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      example:
        id: 1
        name: Some Customer
        address: Some address
        created_at: '2020-02-10 17:32:28'
        updated_at: '2020-02-10 17:32:28'
    CustomerRequestBody:
      description: Create Customer
      type: object
      required:
        - name
        - address
      properties:
        name:
          type: string
        address:
          type: string
      example:
        name: Some Customer
        address: Some address
    CustomersResponse:
      description: List of Customers
      type: array
      items:
        $ref: '#/components/schemas/CustomerSchema'
    OrderSchema:
      description: Order schema
      type: object
      required:
        - id
        - customer_id
        - payment_status
        - status
        - address
        - created_at
        - updated_at
      properties:
        id:
          type: number
        customer_id:
          type: number
        payment_status:
          type: string
        status:
          type: string
        address:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      example:
        id: 1
        customer_id: 1
        payment_status: notPaid
        status: created
        address: some address
        created_at: '2020-02-10 17:32:28'
        updated_at: '2020-02-10 17:32:28'
    OrderRequestBody:
      description: Create Order
      type: object
      required:
        - customer_id
        - payment_status
        - status
      properties:
        customer_id:
          type: number
        payment_status:
          type: string
        status:
          type: string
        address:
          type: string
      example:
        customer_id: 1
        payment_status: notPaid
        status: created
        address: some address
    OrdersResponse:
      description: List of Orders
      type: array
      items:
        $ref: '#/components/schemas/OrderSchema'
...
    CourierOrderSchema:
      description: CourierOrder schema
      type: object
      required:
        - id
        - courier_id
        - order_id
        - created_at
        - updated_at
      properties:
        id:
          type: number
        courier_id:
          type: number
        order_id:
          type: number
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      example:
        id: 1
        courier_id: 1
        order_id: 1
        created_at: '2020-02-10 17:32:28'
        updated_at: '2020-02-10 17:32:28'
    CourierOrderRequestBody:
      description: Create CourierOrder
      type: object
      required:
        - courier_id
        - order_id
      properties:
        courier_id:
          type: number
        order_id:
          type: number
      example:
        courier_id: 1
        order_id: 1
    CourierOrdersResponse:
      description: List of CourierOrders
      type: array
      items:
        $ref: '#/components/schemas/CourierOrderSchema'
    CommonFilter:
      type: object
      properties:
        offset:
          type: integer
          minimum: 0
          example: 0
        limit:
          type: integer
          minimum: 1
          example: 100
        order_by:
          type: object
          properties:
            sort_direction:
              type: string
              enum: [ASC, DESC]
              example: ASC
            field:
              type: string
              example: created_at